Splitted test files into multiple seperate files, to gain overview.
